+++ .templates/default/src/lib/test/test.utils.ts
@@ -2,7 +2,7 @@ import { test } from "ava";
 import * as utils from "../utils";
 
 test( "hello", t => {
-  t.is( utils.hello(), "Hello TSCLI!" );
+  t.is( utils.hello(), "Hello DM-Tools!" );
 } );
 
 test( "goodbye", t => {